Use dtsi template with placeholders for dtsi writer
For better maintainability, it's better to use dtsi template with place
holders for dtsi writer. The writer just need to format the output
string, which is read from the template, with the keys to populate.

Note that the template may involve {} after the text with the desired
keys. To avoid massively escaping those literals, a token cqm_lgm is
used to identify the text, after which {} will be replaced with <<>>.
After formating the output string, they will be replaced back.